pyproject.toml
[project]
name = "skypilot-iceberg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.3",
    "duckdb>=1.2.2",
    "fastparquet>=2024.11.0",
    "gitpython>=3.1.44",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pyiceberg[sql-sqlite]>=0.9.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "supabase>=2.15.0",
]
